Implement IUserInterface on MainForm, and code cleanup
